Update create and push script
Swallow file/directory not found messages

+++ create-and-push-package.bat
@@ -1,8 +1,8 @@
 @echo off
 cd /d %~dp0
-del /q PerfectXL.EPPlus\bin\Release\PerfectXL.EPPlus.*.nupkg
-rd /s/q PerfectXL.EPPlus\bin
-rd /s/q PerfectXL.EPPlus\obj
+if exist PerfectXL.EPPlus\bin\Release\PerfectXL.EPPlus.*.nupkg del /q PerfectXL.EPPlus\bin\Release\PerfectXL.EPPlus.*.nupkg
+if exist PerfectXL.EPPlus\bin rd /s/q PerfectXL.EPPlus\bin
+if exist PerfectXL.EPPlus\obj rd /s/q PerfectXL.EPPlus\obj
 dotnet pack -c Release
 dotnet nuget push PerfectXL.EPPlus\bin\Release\PerfectXL.EPPlus.*.nupkg --api-key q5ZDX89QvzM01aG4F8ZJ --source https://nuget.perfectxl.com/nuget/default
 pause
